requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
defusedxml==0.7.1
Django==3.2.8
django-braces==1.15.0
django-cors-headers==3.10.1
django-filter==21.1
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
httpie==2.6.0
idna==3.3
numbertoletters==0.1.3
Pillow==8.3.2
pisa==3.0.33
Pygments==2.11.2
PyPDF2==1.26.0
PySocks==1.7.1
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.6
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.5
